+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-4491] Refactored permissions model to resolve permissions within the
+  taxonomy tree (i.e. to consider permissions not only in the specific domain
+  being queried, but also in its parent domains). The
+  `Permissions.has_permission?/2` and `Permissions.has_permission?/4` functions
+  now also considers global permissions (the permissions of the default role)
+  before querying session-specific permissions.
+- `CacheCleaner` is no longer a `GenServer`. Instead, use `CacheCleaner.clean/1`
+  (e.g. from a scheduled task)
+
+### Removed
+
+- The `PermissionsConfig` module is no longer used and has been removed
+- The following unused and deprecated functions have been removed:
+  - `DomainCache.id_to_parent_ids_map/0`
+  - `Permissions.get_acls_by_resource_type/2`
+  - `TaxonomyCache.get_descendent_ids/1`
+  - `TaxonomyCache.get_domain_name_to_id_map/0`
+  - `TaxonomyCache.get_domain_external_id_to_id_map/0`
+  - `TaxonomyCache.get_name/1`
+  - `TaxonomyCache.get_parent_ids/1`
+  - `TaxonomyCache.get_root_domain_ids/0`
+
 ## [4.39.0] 2022-03-04
 
 ### Added
@@ -10,7 +36,7 @@
 
 ### Added
 
-- [TD-4481]  New permission `manage_business_concepts_domain`
+- [TD-4481] New permission `manage_business_concepts_domain`
 
 ## [4.38.0] 2022-02-17